Move Theano-dependent code in distributions pkg
All the modules from the distribution package were moved into
garage.theano.
