--- garage/distributions/recurrent_diagonal_gaussian.py
@@ -1,7 +1,0 @@
-import numpy as np
-import theano.tensor as TT
-
-from garage.distributions import DiagonalGaussian
-from garage.distributions import Distribution
-
-RecurrentDiagonalGaussian = DiagonalGaussian
