+++ garage/policies/categorical_mlp_policy.py
@@ -3,13 +3,13 @@ import lasagne.nonlinearities as NL
 import numpy as np
 
 from garage.core import Serializable
-from garage.distributions import Categorical
 from garage.misc import ext
 from garage.misc.overrides import overrides
 from garage.policies import StochasticPolicy
 from garage.spaces import Discrete
 from garage.theano.core import LasagnePowered
 from garage.theano.core import MLP
+from garage.theano.distributions import Categorical
 
 
 class CategoricalMLPPolicy(StochasticPolicy, LasagnePowered):
