+++ garage/policies/gaussian_gru_policy.py
@@ -5,13 +5,13 @@ import numpy as np
 import theano.tensor as TT
 
 from garage.core import Serializable
-from garage.distributions import RecurrentDiagonalGaussian
 from garage.misc import ext
 from garage.misc.overrides import overrides
 from garage.policies import StochasticPolicy
 from garage.theano.core import GRUNetwork
 from garage.theano.core import LasagnePowered
 from garage.theano.core import ParamLayer
+from garage.theano.distributions import RecurrentDiagonalGaussian
 
 
 class GaussianGRUPolicy(StochasticPolicy, LasagnePowered):
