+++ garage/regressors/categorical_mlp_regressor.py
@@ -5,7 +5,6 @@ import theano
 import theano.tensor as TT
 
 from garage.core import Serializable
-from garage.distributions import Categorical
 from garage.misc import ext
 from garage.misc import logger
 from garage.misc import special
@@ -13,6 +12,7 @@ from garage.optimizers import LbfgsOptimizer
 from garage.optimizers import PenaltyLbfgsOptimizer
 from garage.theano.core import LasagnePowered
 from garage.theano.core import MLP
+from garage.theano.distributions import Categorical
 
 NONE = list()
