+++ garage/regressors/gaussian_conv_regressor.py
@@ -6,7 +6,6 @@ import theano
 import theano.tensor as TT
 
 from garage.core import Serializable
-from garage.distributions import DiagonalGaussian
 from garage.misc import logger
 from garage.misc.ext import compile_function
 from garage.misc.ext import iterate_minibatches_generic
@@ -15,6 +14,7 @@ from garage.optimizers import PenaltyLbfgsOptimizer
 from garage.theano.core import ConvNetwork
 from garage.theano.core import LasagnePowered
 from garage.theano.core import ParamLayer
+from garage.theano.distributions import DiagonalGaussian
 
 
 class GaussianConvRegressor(LasagnePowered):
