+++ garage/theano/distributions/__init__.py
@@ -1,0 +1,8 @@
+from garage.theano.distributions.base import Distribution
+from garage.theano.distributions.categorical import Categorical
+from garage.theano.distributions.delta import Delta
+from garage.theano.distributions.diagonal_gaussian import DiagonalGaussian
+from garage.theano.distributions.recurrent_categorical import \
+        RecurrentCategorical
+from garage.theano.distributions.recurrent_diagonal_gaussian import \
+        RecurrentDiagonalGaussian
