+++ garage/theano/distributions/recurrent_diagonal_gaussian.py
@@ -1,0 +1,7 @@
+import numpy as np
+import theano.tensor as TT
+
+from garage.theano.distributions import Distribution
+from garage.theano.distributions import DiagonalGaussian
+
+RecurrentDiagonalGaussian = DiagonalGaussian
